FEC-1736 #comment Fix wrong access to mediaProxy params #time 0.5d
Initial values of mediaProxy are set as full strings in config, e.g.
mediaProxy.mediaPlayFrom is set as key: mediaProxy.mediaPlayFrom, while
the setKdpAttribute saves the values as compound object, e.g.
mediaProxy:{mediaPlayFrom: someValueHere}.
In general we need to align this to be same for all input output
scenarios.