TOG-125 #comment Prevent force ending clip #time 4h
If midroll is set very close to end of entry (if precision is necessary
then if midriff start time is set to be greater or equal to 0.99% of
entry duration :-)) then after a predefined amount of time the player
forces clip done - this is to protect against cases (android) where
video doesn’t trigger ended event.
The above is not good for midriff cases, but only for actual entry
playback so added a check to not force clip done if we are in sequence.