Instead of declaring mw.AdLoader as a static object, allow it to be initalized so that the embedPlayer variable can be set and accessed, allowing creation of an instance of an AdLoader.
Modify the VastParser static object to include the AdLoader object that is calling it, further allowing access to the embedPlayer variable from the loader.

Once the embedPlayer variable is available from within the VastParser, we can call it's evaluate function on any VastAdTagURI, and substitute ad variables when they are declared in there too.