Fix playhead update when seeking in paused state
When seeking in paused state and the used player is flash-osmf HLS we
don’t update the playhead correctly.
Reason is that we are not getting the total bytes from player when
entry first loads and then all percentage calculations are NaN.
The bug was always there probably, but only when we fixed
https://github.com/kaltura/mwEmbed/pull/2890 the issue started to make
problems visible.
FEC-5827