fix(FEC-11256): video controls are not clickable on IOS when captureClickEventForiOS flag is used (#4226)
the issue:
when playing media in IOS and captureClickEventForIOS flag is set to true, the native ios controls are not working.
when captureClickEventForIOS flag is set to true, there is re-writing of the video element. seems like "videoDisplay" div was forgotten and as a result, the hierarchy in DOM was changed.

the solution:
added videoDisplay div before video element, so the hierarchy will be the same as when the flag is disabled.